event file refactor
